Create DCC Feb 4 - Contiguous Array
Link to the question : https://leetcode.com/problems/contiguous-array/
I have used Hashmap<Integer,Integer> for storing sum and index. Initially, put 0,-1 for initializing . Treating 0 as -1 and 1 as 1 so that we can calculate the sum and storing it in the sum variable. Whenever we see a previously encountered sum, we find its corresponding index and calculate the length to see if it is greater than previously stored length. If yes, we update it in result. If it has not been previously encountered, we put it in the hashmap. At last, we return the result.