Create DCC Feb 9 - K-diff pairs in an Array
Link to the question : https://leetcode.com/problems/k-diff-pairs-in-an-array/
First we will build our HashMap. In Map we will keep the no. as a key & value as a count of occurance.
Now, there are 2 cases :-
If k > 0, then in this case we just need to check whether the counter part exists or not. So, if we are iterating 'x + k' in our map, then we can increment our count
If k == 0, then we just need to check if x is more then 1 or not [x > 1] in our map.
But another thing we need to note that here as we iterate from the array after doing counting. Then we will get "1" two times. So, to avoid this instead of iterating over the array, we will iterate over the keyset of this map, which will give us the unique no. i.e. (3,1,4,5).