Create DCC Feb 13 - Subsets
Link to the question : https://leetcode.com/problems/subsets/
Approach : For every element, there are two choices 1. Include it in the arraylist , 2. Do not include it in the arraylist.
The powerset of the array contains 2^n elements where n is the number of elements in an array.

Eg : arr=[10,20,30] 
The powerset will contain 8 lists.
The trick is simple, iterate the array from zero to length(powerset)-1. For each i, convert it into binary. Then create an inner loop to traverse the binary digit. If it has zero, do not include the array at that position and if it contains 1, include the element at that position.
0  - 000 (include nothing i.e. [] )
1 - 001 (include element at third position i.e. [30])
2 - 010 (include element at second position i.e. [20])
3 - 011 (include element at second and third position i.e. [20,30])
4 - 100 ((include element at first position i.e. [10])
5 - 101 (include element at first and third position i.e. [10,30])
6 - 110 (include element at first and second position i.e. [10,20])
7 - 111 (include element at first, second and third position i.e. [10,20,30])
After every step. add it to main list o lists and at the end, return that.

Thanks to Sumeet Sir (Pepcoding) for this detailed explanation.